[change] แก้reload
