+++ components/What/index.js
@@ -38,6 +38,25 @@ const Icon = styled.img`
     padding: .2em;
     width:7em;
   }
+  cursor:pointer;
+  &:hover{
+    background: transparent;
+    -moz-transform: scale(1.2);
+    -webkit-transform: scale(1.2);
+    transform: scale(1.2);
+    border-radius: 0.5em;
+    border: 4px solid transparent;
+    
+  }
+  &:active{
+  background: transparent;
+  -moz-transform: scale(1.3);
+  -webkit-transform: scale(1.3);
+  transform: scale(1.3);
+  border-radius: 1em;
+  border-radius: 0.5em;
+  border: 4px solid transparent;
+  }  
 `
 const P = styled.p`
 font-size:1.2em;
@@ -67,7 +86,7 @@ const IconInfo = [
   { icon: "/static/image/icon-what/java (2).png", text: "Java Programming", t: "Java Programming" },
   { icon: "/static/image/icon-what/IT fun (1).png", text: "IT Fundamentals", t: "IT Fundamentals" },
   { icon: "/static/image/icon-what/network (2).png", text: "Network", t: "Network" },
-  { icon: "/static/image/icon-what/web (2).png", text: "HTML5&CSS", t: "HTML5&CSS" }
+  { icon: "/static/image/icon-what/web (2).png", text: "HTML5&CSS3", t: "HTML5&CSS3" }
 ]
 
 
@@ -116,7 +135,7 @@ const Font = styled.p`
 `
 const index = props => (
 
-  <Bggueng themeColor={props.bgColor.what}>
+  <Bggueng themeColor={props.color.what}>
     <Relative>
       <Moutain src='/static/image/MoutainWho.svg' />
       <div className="container px-3">
